teldir and sms_queue ported and slightly changed. Yet no tested

+++ src/sms_queue.h
@@ -13,18 +13,18 @@
 
 struct SMS_Queque_TypeDef{
     struct SMS_TypeDef{
-        u32 LifeTime;
+        u8 LifeTime;
         u8 TelNum[SMS_TELNUM_LEN];
         u8 *SmsText;
     }List[SMS_QUEUE_MAXSIZE];
-    u32 FirstItem;
-    u32 LastItem;
-    u32 NumItems;
+    u8 FirstItem;
+    u8 LastItem;
+    u8 NumItems;
 };
 
 void SMS_Queue_Init(void);
-void SMS_Queue_Push(u8 *TelNum, const u8 SmsText[], u32 LifeTime);
-u32 SMS_Queue_Pop(u8 *TelNum, u8 **SmsText);
-u32 SMS_Queue_NumItems(void);
+void SMS_Queue_Push(u8 *TelNum, const u8 SmsText[], u8 LifeTime);
+u8 SMS_Queue_Pop(u8 *TelNum, u8 **SmsText);
+u8 SMS_Queue_NumItems(void);
 
 #endif
